extract format timestamp to util + add interaction to list rows
