+++ src/components/layouts/ListLayout.module.css
@@ -1,0 +1,5 @@
+.clickableListLayout {
+  [class~=ant-table-row] {
+    cursor: pointer;
+  }
+}
