+++ src/pages/CatalogsList.tsx
@@ -1,17 +1,20 @@
 import React from 'react';
 import { useListCatalogs } from '../hooks/catalog';
-import ListLayout from '../components/ListLayout';
+import ListLayout from '../components/layouts/ListLayout';
 import { Typography } from 'antd';
+import { formatTimestamp } from '../utils/formatTimestamp';
+import { useNavigate } from 'react-router-dom';
 
 export default function CatalogsList() {
-  const { data } = useListCatalogs();
-
-  if (!data) return null;
+  const { data, isLoading } = useListCatalogs();
+  const navigate = useNavigate();
 
   return (
     <ListLayout
+      loading={isLoading}
       title={<Typography.Title level={2}>Catalogs</Typography.Title>}
-      data={data.catalogs}
+      data={data?.catalogs}
+      onRowClick={(record) => navigate(`/data/${record.name}`)}
       columns={[
         { title: 'Name', dataIndex: 'name', key: 'name', width: '60%' },
         {
@@ -19,15 +22,7 @@ export default function CatalogsList() {
           dataIndex: 'created_at',
           key: 'created_at',
           width: '40%',
-          render: (value) =>
-            new Intl.DateTimeFormat('en-US', {
-              year: 'numeric',
-              month: '2-digit',
-              day: '2-digit',
-              hour: 'numeric',
-              minute: 'numeric',
-              second: 'numeric',
-            }).format(new Date(value)),
+          render: (value) => formatTimestamp(value),
         },
       ]}
     />
