+++ src/utils/formatTimestamp.ts
@@ -1,0 +1,10 @@
+export function formatTimestamp(timestamp: number): string {
+  return new Intl.DateTimeFormat('en-US', {
+    year: 'numeric',
+    month: '2-digit',
+    day: '2-digit',
+    hour: 'numeric',
+    minute: 'numeric',
+    second: 'numeric',
+  }).format(new Date(timestamp));
+}
